Fix unused variable in test_other_frameworks_no_available_plugins.
